Memperbaiki error dan create internal database

+++ app/src/main/java/com/ubaya/a160419046_ubayakost/view/UserDetailFragment.kt
@@ -31,7 +31,7 @@ class UserDetailFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         viewModel = ViewModelProvider(this).get(UserDetailViewModel::class.java)
-        viewModel.fetch(GlobalData.userid)
+        viewModel.fetch(GlobalData.userid.toInt())
 
         observeViewModel()
     }
